requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=10.0.0
opencv-python>=4.8.0
tqdm>=4.65.0
onnx>=1.14.0  # 可选，用于模型导出
PyQt5>=5.15.0  # GUI界面

